Add image to clients
