+++ drivers/hmis/app/graphql/types/base64.rb
@@ -1,0 +1,14 @@
+# app/graphql/types/json_object.rb
+class Types::Base64 < Types::BaseScalar
+  description 'A base64 encoded string'
+
+  def self.coerce_input(input_value, _context)
+    # Comes as a string, so just pass it
+    input_value
+  end
+
+  def self.coerce_result(ruby_value, _context)
+    # Just a string, so just pass it
+    ruby_value
+  end
+end
